Update __init__.py (remove(path), copy())
remove(path): Windows throws a PermissionError if attempting to execute os.unlink(path) on a path with read-only files, like the ones from the Pleasuredome seed after extraction. The test for isdir() lets us bypass this and throw it straight to remove_folder(path), which handles this situation without an error.
Windows workaround also required in copy() function to avoid attempting to use a file handler on a folder